labs/pyproject.toml
[project]
name = "ibm-quantum-computing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pylatexenc>=2.10",
    "python-dotenv>=1.0.0",
    "qiskit[all]~=2.3.0",
    "qiskit-ibm-runtime~=0.43.1",
]
